+++ terraform/lib/auth/.gitignore
