+++ terraform/lib/auth/deployment_package.zip
