+++ terraform/lib/auth/index.js
