+++ terraform/lib/auth/package.json
